Implement support for large alignments
* Change:

Implement support for alignment above LARGE_PAGE_SIZE, but still strictly
less than HUGE_PAGE_SIZE.

En-passant, fix the LLAllocator to comply with Rust Layout semantics,
in Layout the size is not necessary a multiple of the alignment.

* Motivation:

It would be rather hypocrite to make use of large alignments to build
an allocator which does not allow them.