Add README and Design document
* Changes:

- Add README.
- Add Design document.

* Motivation:

A README presenting the goals, non-goals, and status of the project is
an important piece of information for would-be users, allowing them to
understand whether the project suits their own needs, or not.

A Design document presenting the rationale for a variety of decisions
is useful for both maintainers, and curious souls. Maintainers will
appreciate understanding the rationale behind certain decisions, and
therefore whether their "better" solution does not accidentally
sacrifice important properties. Curious souls may find nuggets of
wisdom.