Modernize Part 1: NonNull, Blocks.
* Changes:

- Use NonNull rather than raw pointers.
- Rename Cells to Blocks.
- Split blocks.rs into multiple files.

* Motivation:

NonNull and Option<NonNull> make it much more explicit in the type
system which pointers are expected to be potentially null, and which are
not. Option also offers monadic combinators.

Cell is a painful name to the conflict with core::cell, so Block it is.

And more large file splitting is necessary in internals, there's still a
few really massive files there.