Introduce AtomicStack, use it in ThreadLocalManager.
* Changes:

- Introduce AtomicStack.
- Switch ThreadLocalManager from AtomicBlockForeignStack to AtomicStack.
- Remove AtomicBlockForeignStack.

* Motivation:

AtomicStack is a fresh take on a wait-free intrusively-linked stack,
which seeks to work around the typical ABA issue faced by a naive
implementation using a merkle chain.

That is, the typical ABA of linked stacks occurs in pop implementations:

- Thread A reads head, and its next field.
- Thread B pops head, pops next, pushes back head.
- Thread proceeds with the CAS of head with next, even though next is
  now popped -- and used externally.

This ABA accident only occurs in case of "nigh-immediate" re-push, which
may be rare in a number of usecases -- such as when pops and pushes are
rare to start with.

This implementation improves _slightly_ on it by using a merkle chain.
Specifically, considering a stack node, the `next` field contains not a
pointer, but instead a pointer + a hash of the pointee -- recursively.
The two are mashed together in 64 bits, with the "unused" pointer bits
(platform limitations + alignment) used for the hash.

The result is that in the case above, if B re-pushes `head` with a new
`next` field, then the stored `top` value will contain the `hash` of the
_new_ next field, and unless they match the `hash` of the _old_ next
field the CAS exchange will fail.

To be clear, the CAS exchange is NOT guaranteed to fail, however, as
long as the hash is fairly good, and as many bits as possible are used
for it, then the chances that of the hash bits being equal are very low.
In the case of a scenario unlikely to happen to start with, this should
be fairly good.

Note: another possibility would be to keep a _counter_ somewhere, and
mix in the _counter_ bits. This would however double the stack space.